fix: sidebar links opening modal
