+++ src/models/ui.ts
@@ -11,3 +11,14 @@ export type QatButtonProps = {
   color: string;
   icon: IconDefinition;
 };
+
+export type SidebarProps = {
+  openModal: (arg0: boolean) => void;
+};
+
+export type NavProps = {
+  link?: string;
+  icon: IconDefinition;
+  name: string;
+  openModal?: (arg0: boolean) => void;
+};
